This is the index file.
Since i usually create a separate folder inside dashboard of XAMPP for each project, The index files for each project reside inside dashboard only. Which in turn isolates the index file from rest of the project in terms of pushing it to a git repository. Therefore i am pushing index file separately as it is not in my project(git) folder of this project.